Move modules from components to containers
- AdminPageLayout
- HeroPageLayout
